+++ miopengemm/include/miopengemm/bundle.hpp
@@ -20,13 +20,13 @@ class Bundle
 {
   public:
   const std::vector<KernelString>          v_tgks;
-  const std::vector<std::vector<unsigned>> v_wait_indices;
+  const std::vector<std::vector<size_t>> v_wait_indices;
 
   derivedparams::DerivedParams dp;
 
   /* TODO : when is std::move needed. */
   Bundle(std::vector<KernelString>&&          v_tgks_,
-         std::vector<std::vector<unsigned>>&& v_wait_indices_,
+         std::vector<std::vector<size_t>>&& v_wait_indices_,
          derivedparams::DerivedParams&&       dp_)
     : v_tgks(std::move(v_tgks_)), v_wait_indices(std::move(v_wait_indices_)), dp(std::move(dp_))
   {
